[PBA-5530] Change integration instructions
I tested adding the Pulse_tvOS.framework into the app and it only works if you add it to the Embedded Binaries too, not just only Linked Frameworks and Libraries. I’m not adding the Pulse_tvOS.framework here because we still ask people to download it from the Ooyala site.